# lib/api-zod/src/generated/types/messageThread.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voom Car Rental API
 * OpenAPI spec version: 0.1.0
 */

export interface MessageThread {
  userId: number;
  username: string;
  fullName?: string | null;
  profilePicture?: string | null;
  lastMessage: string;
  lastMessageAt?: Date | null;
  unreadCount: number;
}