# lib/api-zod/src/generated/types/paymentResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voom Car Rental API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentResponse {
  paymentId: number;
  amount: number;
  currency: string;
  status: string;
  message?: string | null;
}